+++ crates/ruma-identity-service-api/src/lookup/lookup_3pid.rs
@@ -11,16 +11,16 @@ pub mod v2 {
 
     use ruma_common::{
         OwnedUserId,
-        api::{auth_scheme::AccessToken, request, response},
+        api::{request, response},
         metadata,
     };
 
-    use crate::lookup::IdentifierHashingAlgorithm;
+    use crate::{IdentityServiceToken, lookup::IdentifierHashingAlgorithm};
 
     metadata! {
         method: POST,
         rate_limited: false,
-        authentication: AccessToken,
+        authentication: IdentityServiceToken,
         history: {
             1.0 => "/_matrix/identity/v2/lookup",
         }