+++ crates/ruma-common/CHANGELOG.md
@@ -28,8 +28,12 @@ Improvements:
   dropped.
 - `UserProfile::merge` has been renamed to `UserProfile::apply`.
 - Add experimental support for [MSC4495] (Selective Presence).
+- Add `required_scopes` function to `Metadata` and accompanying `required_scopes` syntax
+  to `metadata!` macro, to allow request structs to define what OAuth 2.0 scopes they require.
+- Add unstable support for [MSC4484] "Server Administration OAuth Scope".
 
 [MSC4438]: https://github.com/matrix-org/matrix-spec-proposals/pull/4438
+[MSC4484]: https://github.com/matrix-org/matrix-spec-proposals/pull/4484
 [MSC4495]: https://github.com/matrix-org/matrix-spec-proposals/pull/4495
 
 ## 0.19.0